Full storage rework to allow linear data iteration

+++ Assets/Game/Scripts/UnityWorldTest.cs
@@ -74,7 +74,7 @@ namespace ECS
 
 			foreach (var ent in entities)
 			{
-				world.Add(ent, new Armor() { Defense = ent.Index });
+				world.Add(ent, new Armor() { Defense = (int)ent.Index });
 			}
 
 			Debug.Log("defense = " + world.GetComponent<Armor>(entity).Defense);
